interpolate: move Point into interpolate so WASM build compiles
interpolate.rs is gated on any(cli, wasm) but imported Point from
timeseries, which is cli-only. Building with --no-default-features
--features wasm therefore failed to compile, violating the non-negotiable
in CLAUDE.md that the library must build under wasm alone.

Move Point, its Eq/PartialEq/Hash impls, and the TOLERANCE constant it
depends on into interpolate.rs (its conceptual home — interpolation
input). Re-export from timeseries.rs so existing CLI call sites keep
resolving crate::timeseries::Point.

Also add required-features = ["cli"] to the fatigue binary so
cargo build --features wasm without cli no longer tries to compile the
cli-gated main().
